# com.hrrm.famoney.application.api.accounts/src/main/resources/openapi.json
{
  "openapi" : "3.0.1",
  "paths" : {
    "/accounts" : {
      "get" : {
        "tags" : [ "accounts" ],
        "operationId" : "getAllAccounts",
        "parameters" : [ {
          "name" : "tags",
          "in" : "query",
          "schema" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A list of all accounts",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Account"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/accounts/tags" : {
      "get" : {
        "tags" : [ "accounts" ],
        "operationId" : "getAllAccountTags",
        "responses" : {
          "200" : {
            "description" : "A list of all account tags",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{accountId}/movements" : {
      "get" : {
        "tags" : [ "accounts" ],
        "operationId" : "getAccountMovements",
        "parameters" : [ {
          "name" : "accountId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "continue",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "count",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A list of account movements of specified account",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AccountMovement"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Account" : {
        "required" : [ "name" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "name" : {
            "type" : "string"
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "AccountMovement" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "accountLink" : {
            "type" : "string"
          },
          "movementGroupId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "amount" : {
            "type" : "number",
            "format" : "double"
          }
        }
      }
    }
  }
}